chore: add spawn broadcast

+++ src/action/placement.c
@@ -1,6 +1,6 @@
 #include "./placement.h"
 
-bool place_bomb(t_map *map, t_bomb *bomb, size_t x, size_t y)
+bool place_bomb(t_map *map, t_bomb *bomb, int x, int y)
 {
     if (map->matrix[y][x].bomb != NULL) {
         return false;
@@ -9,7 +9,7 @@ bool place_bomb(t_map *map, t_bomb *bomb, size_t x, size_t y)
     return true;
 }
 
-bool place_bomberman(t_map *map, t_bomberman *bomberman, size_t x, size_t y)
+bool place_bomberman(t_map *map, t_bomberman *bomberman, int x, int y)
 {
     if (map->matrix[y][x].bomberman != NULL) {
         return false;
