+++ src/render/bomb.h
@@ -4,4 +4,4 @@
 #include <SDL2/SDL.h>
 
 SDL_Texture *load_bomb(SDL_Renderer *renderer);
-void        render_bomb(SDL_Renderer *renderer, SDL_Texture *texture, size_t x, size_t y);
+void        render_bomb(SDL_Renderer *renderer, SDL_Texture *texture, int x, int y);
