+++ src/render/bomberman.c
@@ -1,6 +1,6 @@
 #include "./bomberman.h"
 
-void            render_bomberman(SDL_Renderer *renderer, SDL_Texture *texture, size_t x, size_t y)
+void            render_bomberman(SDL_Renderer *renderer, SDL_Texture *texture, int x, int y)
 {
     (void) texture;
