+++ src/render/button.h
@@ -3,4 +3,4 @@
 #include <stdlib.h>
 #include <SDL2/SDL.h>
 
-void render_button(SDL_Renderer *renderer, SDL_Texture *texture, size_t x, size_t y);
+void render_button(SDL_Renderer *renderer, SDL_Texture *texture, int x, int y);
